bot/pyproject.toml
[project]
name = "pal-300-bracket-discord-bot"
version = "0.1.0"
description = "Discord bot for the Podcast About List 300 episode tournament"
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
	"audioop-lts>=0.2.1",
	"discord-py>=2.4.0",
	"gitpython>=3.1.44",
	"python-dotenv>=1.0.1",
]

[tool.uv]
required-version = ">=0.5.20"

[tool.ruff]
# Give us a lil bit more line length
line-length = 96
indent-width = 2

[tool.ruff.format]
# yum yum yum i love using tabs
indent-style = "tab"

[tool.ruff.lint]
# laughing all the way to the bank.
# holy cow just switch em all on and let 'er rip bud
select = [
	"E", "F", "W", "I", "N", "UP", "YTT", "ANN", "ASYNC", "S", "BLE", "FBT", "B",
	"A", "COM", "C4", "EM", "EXE", "ISC", "ICN", "LOG", "G", "INP", "PIE", "T20",
	"Q", "RSE", "RET", "SLF", "SIM", "TID", "ARG", "TD", "PL", "TRY", "PERF",
	"FURB", "RUF"
]
ignore = [
	"W191",   # you know a spacelet hates to see a tabchad like me coming
	"TD003",  # don't require issue links for TODO messages
	"COM812", # conflicts with `ruff format`
]

[dependency-groups]
dev = [
	"ruff>=0.9.1",
]
